+++ src/BetterRead.Shared.Domain/Search/Context.cs
@@ -1,0 +1,14 @@
+using Newtonsoft.Json;
+
+namespace BetterRead.Shared.Domain.Search
+{
+    public class Context
+    {
+        [JsonProperty("title")]
+        public string Title { get; set; }
+
+        [JsonProperty("total_results")]
+        [JsonConverter(typeof(ParseStringConverter))]
+        public long TotalResults { get; set; }
+    }
+}
