+++ src/BetterRead.Shared.Domain/Search/CseImage.cs
@@ -1,0 +1,11 @@
+using System;
+using Newtonsoft.Json;
+
+namespace BetterRead.Shared.Domain.Search
+{
+    public class CseImage
+    {
+        [JsonProperty("src")]
+        public Uri Src { get; set; }
+    }
+}
