+++ src/BetterRead.Shared.Domain/Search/FindMorOnGoogle.cs
@@ -1,0 +1,11 @@
+using System;
+using Newtonsoft.Json;
+
+namespace BetterRead.Shared.Domain.Search
+{
+    public class FindMoreOnGoogle
+    {
+        [JsonProperty("url")]
+        public Uri Url { get; set; }
+    }
+}
