+++ src/BetterRead.Shared.Domain/Search/Page.cs
@@ -1,0 +1,14 @@
+using Newtonsoft.Json;
+
+namespace BetterRead.Shared.Domain.Search
+{
+    public class Page
+    {
+        [JsonProperty("label")]
+        public long Label { get; set; }
+
+        [JsonProperty("start")]
+        [JsonConverter(typeof(ParseStringConverter))]
+        public long Start { get; set; }
+    }
+}
